@@ -20,21 +20,22 @@
     "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "howler": "^2.0.8",
     "@types/howler": "^2.0.5",
     "@types/node": "^6.0.97",
     "angular2-template-loader": "^0.6.0",
     "awesome-typescript-loader": "^3.0.4",
+    "copy-webpack-plugin": "^4.3.1",
     "css-loader": "^0.26.1",
     "extract-text-webpack-plugin": "2.0.0-beta.5",
     "file-loader": "^0.9.0",
+    "howler": "^2.0.8",
     "html-loader": "^0.4.3",
     "html-webpack-plugin": "^2.16.1",
     "null-loader": "^0.1.1",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.5.2",
     "style-loader": "^0.13.1",
-    "typescript": "~2.3.1",
+    "typescript": "^2.4.1",
     "webpack": "2.2.1",
     "webpack-dev-server": "2.4.1",
     "webpack-merge": "^3.0.0"
